pyproject.toml
[project]
name = "BDGNUPET1"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.12"
dependencies = [
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
sqlalchemy = ">=2.0.23"
pymysql = ">=1.1.0"
python-dotenv = ">=1.0.0"